Fix analytics/Sentry always enabled in development mode
The AnalyticsService.IsEnabled logic was:
  (AnalyticsEnabled && IsProduction) || IsDevelopment

This caused analytics to ALWAYS be enabled in development mode,
regardless of the FIGHTARR__ANALYTICS_ENABLED config setting.

Changed to:
  AnalyticsEnabled && (IsProduction || IsDevelopment)

Now analytics respects the config setting in both production and
development modes.

This fixes the UI hanging on load due to failed DNS lookups to
sentry.fightarr.tv (ERR_NAME_NOT_RESOLVED) when analytics is
disabled via config but forced on by development mode.