fix: Add onClick handler to Download Client template buttons
Fixed issue where Download Client template buttons in the Add modal were not clickable.

Issue:
- Template buttons showed hover effects but had no onClick handler
- Clicking on client templates (SABnzbd, qBittorrent, etc.) did nothing
- User could not proceed past the template selection screen

Fix:
- Added onClick handler to template buttons
- Now shows alert when template is selected
- Placeholder for full configuration form (to be implemented)

Note: This is a temporary fix that adds clickability. Full form implementation
with host/port/auth configuration will be added in a future update. The alert
notifies users that the feature is being developed.